+++ PersistencePlugin/src/com/elmakers/mine/bukkit/plugins/persistence/dao/CommandData.java
@@ -1,8 +1,8 @@
 package com.elmakers.mine.bukkit.plugins.persistence.dao;
 
+import java.util.List;
+
 import org.bukkit.command.Command;
-import org.bukkit.plugin.Plugin;
-import org.bukkit.plugin.PluginDescriptionFile;
 
 import com.elmakers.mine.bukkit.plugins.persistence.annotation.Persist;
 import com.elmakers.mine.bukkit.plugins.persistence.annotation.PersistClass;
@@ -15,7 +15,7 @@ import com.elmakers.mine.bukkit.plugins.persistence.annotation.PersistClass;
  */
 @PersistClass(name="command", schema="global")
 public class CommandData
-{
+{	
 	/**
 	 * The default constructor, used by Persistence to create instances.
 	 */
@@ -33,12 +33,28 @@ public class CommandData
 	 * @param id The command id
 	 * @param command The command string, or alias
 	 */
-	public CommandData(Plugin plugin, String id, String command)
+	public CommandData(PluginData plugin, String command)
 	{
-		PluginDescriptionFile pdfFile = plugin.getDescription();
-		this.pluginId = pdfFile.getName();
-		this.id = id;
+		this.plugin = plugin;
 		this.command = command;
+	}
+	
+	/**
+	 * Add a command to this command as a sub-command.
+	 * 
+	 * Sub-commands are activated using parameters. So:
+	 *
+	 * /persist list global.player.NathanWolf
+	 * 
+	 * Consists of the main Command "persist", one sub-command "list",
+	 * and one parameter "global.player.NathanWolf".
+	 * 
+	 * @param command The command to add as a sub-command of this one
+	 */
+	public void addSubCommand(CommandData command)
+	{
+		children.add(command);
+		command.setParent(this);
 	}
 	
 	/**
@@ -69,13 +85,13 @@ public class CommandData
 		return "/" + command;
 	}
 	
-	@Persist(id=true)
-	public String getId()
+	@Persist(id=true, auto=true)
+	public int getId()
 	{
 		return id;
 	}
 	
-	public void setId(String id)
+	public void setId(int id)
 	{
 		this.id = id;
 	}
@@ -92,17 +108,41 @@ public class CommandData
 	}
 
 	@Persist
-	public void setPluginId(String pluginId)
+	public void setPlugin(PluginData plugin)
 	{
-		this.pluginId = pluginId;
+		this.plugin = plugin;
 	}
 
-	public String getPluginId()
+	public PluginData getPlugin()
 	{
-		return pluginId;
+		return plugin;
 	}
 
-	private String id;
-	private String command;
-	private String pluginId;
+	@Persist
+	public void setParent(CommandData parent)
+	{
+		this.parent = parent;
+	}
+
+	public CommandData getParent()
+	{
+		return parent;
+	}
+	
+	@Persist
+	public void setChildren(List<CommandData> children)
+	{
+		this.children = children;
+	}
+
+	public List<CommandData> getChildren()
+	{
+		return children;
+	}
+
+	private int					id;
+	private String				command;
+	private PluginData			plugin;
+	private CommandData			parent;
+	private List<CommandData>	children;
 }